Guest Mode COMPLETE
First phase of the project done.

This version provides a basic usage of the canvas to draw kanjis, predictions appear below.

If needed, clicking on any prediction will generate information about the kanji: a svg image (animation on click) and reading/meaning + related vocabulary.

The form to log in with a Wanikani API does not work in this version, though an empty function has been set up to manage this in the second phase.